go-microservice-mongo
This project implements microservices architecture using Golang and MongoDB. It prioritizes modularity and testability, breaking the system into self-contained modules and emphasizing writing tests. Golang provides efficiency and simplicity, while MongoDB offers flexible and scalable data storage. The result is a well-structured and reliable system that is easy to maintain and scale.